Continue renaming filters
