+++ api/graphql/loaders.py
@@ -10,7 +10,7 @@ from fastapi import Request
 from strawberry.dataloader import DataLoader
 
 from api.utils import get_projectless_db_connection, group_by
-from db.python.db_filters import GenericFilter, get_hashable_value
+from db.python.filters import GenericFilter, get_hashable_value
 from db.python.layers import (
     AnalysisLayer,
     AssayLayer,
