+++ db/python/filters/__init__.py
@@ -1,6 +1,6 @@
 # This file is used to import all the filters that are used in the application
 
-from db.python.db_filters.generic import (
+from db.python.filters.generic import (
     GenericFilter,
     GenericFilterModel,
     GenericMetaFilter,
