+++ db/python/layers/cohort.py
@@ -1,5 +1,5 @@
 from db.python.connect import Connection
-from db.python.db_filters import GenericFilter
+from db.python.filters import GenericFilter
 from db.python.layers.base import BaseLayer
 from db.python.layers.sequencing_group import SequencingGroupLayer
 from db.python.tables.cohort import CohortFilter, CohortTable, CohortTemplateFilter
