+++ db/python/layers/family.py
@@ -2,7 +2,7 @@
 
 from api.utils import group_by
 from db.python.connect import Connection
-from db.python.db_filters import GenericFilter
+from db.python.filters import GenericFilter
 from db.python.layers.base import BaseLayer
 from db.python.layers.participant import ParticipantLayer
 from db.python.tables.family import FamilyFilter, FamilyTable
