+++ db/python/layers/ourdna/dashboard.py
@@ -7,7 +7,7 @@ from math import ceil
 from typing import Any
 
 from db.python.connect import Connection
-from db.python.db_filters import GenericFilter
+from db.python.filters import GenericFilter
 from db.python.layers.base import BaseLayer
 from db.python.layers.participant import ParticipantLayer
 from db.python.layers.sample import SampleLayer
