+++ db/python/tables/analysis.py
@@ -4,7 +4,7 @@ import datetime
 from collections import defaultdict
 from typing import Any, Dict, List, Optional, Set, Tuple
 
-from db.python.db_filters import GenericFilter, GenericFilterModel, GenericMetaFilter
+from db.python.filters import GenericFilter, GenericFilterModel, GenericMetaFilter
 from db.python.tables.base import DbBase
 from db.python.utils import NotFoundError, to_db_json
 from models.enums import AnalysisStatus
