+++ db/python/tables/assay.py
@@ -4,7 +4,7 @@ import re
 from collections import defaultdict
 from typing import Any, NamedTuple
 
-from db.python.db_filters import GenericFilter, GenericFilterModel, GenericMetaFilter
+from db.python.filters import GenericFilter, GenericFilterModel, GenericMetaFilter
 from db.python.tables.base import DbBase
 from db.python.utils import NoOpAenter, NotFoundError, to_db_json
 from models.models.assay import AssayId, AssayInternal
