+++ db/python/tables/family.py
@@ -2,7 +2,7 @@ import dataclasses
 from collections import defaultdict
 from typing import Any, Dict, List, Optional, Set
 
-from db.python.db_filters import GenericFilter, GenericFilterModel
+from db.python.filters import GenericFilter, GenericFilterModel
 from db.python.tables.base import DbBase
 from db.python.utils import NotFoundError, escape_like_term
 from models.models.family import FamilyInternal
