+++ test/test_assay.py
@@ -3,8 +3,8 @@ from test.testbase import DbIsolatedTest, run_as_sync
 
 from pymysql.err import IntegrityError
 
-from db.python.db_filters import GenericFilter
 from db.python.enum_tables import AssayTypeTable
+from db.python.filters import GenericFilter
 from db.python.layers.assay import AssayLayer
 from db.python.layers.sample import SampleLayer
 from db.python.tables.assay import AssayFilter, AssayTable
