+++ test/test_cohort.py
@@ -3,7 +3,7 @@ from test.testbase import DbIsolatedTest, run_as_sync
 
 from pymysql.err import IntegrityError
 
-from db.python.db_filters import GenericFilter
+from db.python.filters import GenericFilter
 from db.python.layers import CohortLayer, SampleLayer
 from db.python.tables.cohort import CohortFilter
 from models.models import (
