+++ test/test_participant.py
@@ -1,6 +1,6 @@
 from test.testbase import DbIsolatedTest, run_as_sync
 
-from db.python.db_filters import GenericFilter
+from db.python.filters import GenericFilter
 from db.python.layers.participant import ParticipantLayer
 from db.python.tables.participant import ParticipantFilter
 from models.base import PRIMARY_EXTERNAL_ORG
