+++ test/test_sample.py
@@ -1,7 +1,7 @@
 from test.testbase import DbIsolatedTest, run_as_sync
 
-from db.python.db_filters.generic import GenericFilter
-from db.python.db_filters.sample import SampleFilter
+from db.python.filters.generic import GenericFilter
+from db.python.filters.sample import SampleFilter
 from db.python.layers.sample import SampleLayer
 from models.models import PRIMARY_EXTERNAL_ORG, SampleUpsertInternal
